Source rejig, refactor POSIX to support changes
