+++ workspace/metsrv/metsrv.vcxproj
@@ -655,7 +655,7 @@ copy /y "$(TargetDir)$(TargetFileName)" "$(ProjectDir)..\..\output\$(PlatformSho
     </ClCompile>
     <ClCompile Include="..\..\source\server\win\remote_dispatch.c" />
     <ClCompile Include="..\..\source\server\remote_dispatch_common.c" />
-    <ClCompile Include="..\..\source\server\server_setup.c" />
+    <ClCompile Include="..\..\source\server\server_setup_win.c" />
     <ClCompile Include="..\..\source\server\win\server_setup_winhttp.c">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='r7_debug|Win32'">NotUsing</PrecompiledHeader>
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='r7_debug|Win32'">
